Clicking add to card adds to localstorage and takes you to cart page

+++ eslint.config.js
@@ -16,7 +16,7 @@ export default defineConfig([
     rules: {
       "no-unused-vars": "warn",
       "no-undef": "warn",
-      quotes: ["error", "double", { allowTemplateLiterals: true }],
+      quotes: ["error", "single", { allowTemplateLiterals: true }],
       "no-console": "warn",
     },
   },
